Updated Rosie Demo
Made a new function that unifies the top-level demo for Rosie across different base types.  So far this only works for the omni-base, so things still have to be fixed in the setup for the diff-drive, and a mecanum drive still needs to be implemented.  Added gains for the arm and the base.