Change path in IK example (#62)
On the basic IK example, it traces a straight line from [0.5, -0.2, 0] to [-0.5, -0.2, 0].  This works OK on the 3 dof arm, but points that are passed through in the vicinity of [0.2, -0.2, 0] tend to be a little collision prone with a 6 DOF arm, given the wrist and the gas spring.  I found a trajectory of [0.5, -0.2, 0] to [0.5, 0.2, 0] to work better here, and I think it would work well on the 3 DOF arm as well.  It also seems to center better in the default workspace (so an elbow up configuration with base angle "0" is in the middle).

Thoughts?