Adding Mecanum Base Functionality to Rosie MATLAB Demo (#113)
* basics of a mecanum base code working

This is just some basics to get the mecanum base working in MATLAB for Chris to base stuff on for the ROS example. Will eventually make a version that gets implemented into the rosie demo.

* Added Mecanum support to Rosie demo

Based this off of the ROS and teleop taxi work we did for Ericsson. We have already shipped the Ericsson base so I did not get a chance to test this, but I did some simulated tests making sure that the outputs from the wheelVelocityMatrix are correct.

* update to naming conventions for mobile base wheels